requirements.txt
pytest==8.3.5
flask==3.1.0
flask_sqlalchemy==3.1.1
Flask-WTF==1.2.2
Flask-Login==0.6.3
WTForms==3.2.1
gunicorn==23.0.0
stripe==12.0.1
python-dotenv==1.1.0
